Updated DTO

+++ src/main/java/io/github/lucas_monteiro28/TransactionAPI/dto/response/UserResponseDTO.java
@@ -2,13 +2,15 @@ package io.github.lucas_monteiro28.TransactionAPI.dto.response;
 
 import io.github.lucas_monteiro28.TransactionAPI.model.enums.Role;
 
+import java.math.BigDecimal;
+
 public record UserResponseDTO(
 
         Long id,
         String name,
         String email,
         Role role,
-        Double balance
+        BigDecimal balance
 
 ) {
 }
